Rearranged the files.
Dispatched all the background functions in designated .js files in the js/utils/ folder.
Only kept the init and listener function in the background.js.
All the js files are listed in the right order in the manifest.
New service-worker.js is now only an importScript copying the manifest imports for the background.